Revert "Test cache issues"
