+++ src/components/QuickstartOverview.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { graphql } from 'gatsby';
 import { css } from '@emotion/react';
 import { quickstart } from '../types';
 import Markdown from './Markdown';
@@ -65,5 +66,11 @@ const QuickstartOverview = ({ quickstart }) => {
 QuickstartOverview.propTypes = {
   quickstart: quickstart.isRequired,
 };
+
+export const fragmentQuery = graphql`
+  fragment QuickstartOverview_quickstart on Quickstarts {
+    description
+  }
+`;
 
 export default QuickstartOverview;
